Add schema for owner form - rename file to business-identity-schema
