+++ packages/webcomponents/src/components/business-forms/schemas/business-identity-schema.ts
@@ -13,3 +13,16 @@ export const representativeSchema = object({
   identification_number: string(),
   address: identityAddressSchema,
 });
+
+export const ownerSchema = object({
+  name: string().required('Enter owner name'),
+  email: string()
+    .email('Enter valid owner email')
+    .required('Enter owner email'),
+  phone: string().required('Enter owner phone number'),
+  dob_day: string().required('Enter owner birth day'),
+  dob_month: string().required('Enter owner birth month'),
+  dob_year: string().required('Enter owner birth year'),
+  identification_number: string(),
+  address: identityAddressSchema,
+});
